Move package jsettlers.graphics.reader to jsettlers.graphics.image.reader
