+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/TorsoImage.java
@@ -14,7 +14,7 @@
  *******************************************************************************/
 package jsettlers.graphics.image;
 
-import jsettlers.graphics.reader.ImageMetadata;
+import jsettlers.graphics.image.reader.ImageMetadata;
 
 /**
  * This is a torso: An image that is colored afterwards.
