+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/DatBitmapReader.java
@@ -1,5 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2015
+/*
+ * Copyright (c) 2015 - 2018
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -11,15 +11,15 @@
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
- *******************************************************************************/
-package jsettlers.graphics.reader;
+ */
+package jsettlers.graphics.image.reader;
 
 import java.io.IOException;
 
 import jsettlers.graphics.image.Image;
-import jsettlers.graphics.reader.bytereader.ByteReader;
-import jsettlers.graphics.reader.translator.DatBitmapTranslator;
-import jsettlers.graphics.reader.translator.HeaderType;
+import jsettlers.graphics.image.reader.bytereader.ByteReader;
+import jsettlers.graphics.image.reader.translator.HeaderType;
+import jsettlers.graphics.image.reader.translator.DatBitmapTranslator;
 
 /**
  * This class is capable of reading an image from the given stram.
