+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/DatFileReader.java
@@ -1,5 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2015
+/*
+ * Copyright (c) 2015 - 2018
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -11,30 +11,22 @@
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
- *******************************************************************************/
-package jsettlers.graphics.image;
-
-import jsettlers.common.images.EImageLinkType;
-import jsettlers.graphics.image.reader.ImageMetadata;
-
-/**
- * This is a gui image.
- * 
- * @author Michael Zangl
- * @see EImageLinkType#GUI
  */
-public class GuiImage extends SingleImage {
 
-	/**
-	 * Creates a new GUI image.
-	 * 
-	 * @param metadata
-	 *            The image meta data
-	 * @param data
-	 *            The image data.
-	 */
-	public GuiImage(ImageMetadata metadata, short[] data) {
-		super(metadata, data);
-	}
+package jsettlers.graphics.image.reader;
+
+import java.io.IOException;
+
+import jsettlers.graphics.image.LandscapeImage;
+import jsettlers.graphics.image.reader.bytereader.ByteReader;
+import jsettlers.graphics.image.reader.translator.DatBitmapTranslator;
+
+public interface DatFileReader extends DatFileSet {
+
+	DatBitmapTranslator<LandscapeImage> getLandscapeTranslator();
+
+	ByteReader getReaderForLandscape(int index) throws IOException;
+
+	void generateImageMap(int width, int height, int[] sequences, String id) throws IOException;
 
 }
