+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/ImageArrayProvider.java
@@ -1,5 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2015
+/*
+ * Copyright (c) 2015 - 2018
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -11,42 +11,34 @@
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
- *******************************************************************************/
-package jsettlers.graphics.reader.translator;
+ */
+package jsettlers.graphics.image.reader;
 
 import java.io.IOException;
 
-import jsettlers.graphics.image.TorsoImage;
-import jsettlers.graphics.reader.ImageMetadata;
-import jsettlers.graphics.reader.bytereader.ByteReader;
-
 /**
- * This class reads the torso image. That image is always a grayscale image.
+ * This class gives us a place to store our image
  * 
- * @author Michael Zangl
+ * It is not thread safe.
+ * 
+ * It internally holds a write pointer, and may be backed by an array or a buffer.
+ * 
+ * startImage ist always called first, then writeLine for each line.
+ * 
+ * @author michael
  *
  */
-public class TorsoTranslator implements DatBitmapTranslator<TorsoImage> {
-	private static final int TORSO_BITS = 0x1f;
+public interface ImageArrayProvider {
+	/**
+	 * Starts a new image.
+	 * 
+	 * @param width
+	 *            May be 0!
+	 * @param height
+	 *            May be 0!
+	 * @throws IOException
+	 */
+	void startImage(int width, int height) throws IOException;
 
-	@Override
-	public short getTransparentColor() {
-		return 0x00;
-	}
-
-	@Override
-	public short readUntransparentColor(ByteReader reader) throws IOException {
-		int read = reader.read8() & TORSO_BITS; // only 5 bit.
-		return (short) (read << 11 | read << 6 | read << 1 | 0x01);
-	}
-
-	@Override
-	public HeaderType getHeaderType() {
-		return HeaderType.DISPLACED;
-	}
-
-	@Override
-	public TorsoImage createImage(ImageMetadata metadata, short[] array) {
-		return new TorsoImage(metadata, array);
-	}
+	void writeLine(short[] data, int length) throws IOException;
 }
