+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/custom/graphics/DatFileIndexUtils.java
@@ -1,5 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2015
+/*
+ * Copyright (c) 2018
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -11,34 +11,40 @@
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
- *******************************************************************************/
-package jsettlers.graphics.reader;
+ */
 
-import java.io.IOException;
+package jsettlers.graphics.image.reader.custom.graphics;
+
+import jsettlers.graphics.map.draw.ImageProvider;
+import jsettlers.graphics.image.reader.DatFileReader;
 
 /**
- * This class gives us a place to store our image
- * 
- * It is not thread safe.
- * 
- * It internally holds a write pointer, and may be backed by an array or a buffer.
- * 
- * startImage ist always called first, then writeLine for each line.
- * 
- * @author michael
+ * Different settlers versions use different image indexes.
+ * <p>
+ * In the game, we always reference the image indexes of the gold edition.
+ * <p>
+ * This utility class holds the information about which index remapping to use for which file.
  *
+ * @author michael
  */
-public interface ImageArrayProvider {
-	/**
-	 * Starts a new image.
-	 * 
-	 * @param width
-	 *            May be 0!
-	 * @param height
-	 *            May be 0!
-	 * @throws IOException
-	 */
-	void startImage(int width, int height) throws IOException;
+public class DatFileIndexUtils {
 
-	void writeLine(short[] data, int length) throws IOException;
+	private DatFileIndexUtils() {
+	}
+
+	/**
+	 * Returns a reader that uses gold version indexes
+	 *
+	 * @param fileIndex
+	 * @param reader
+	 * 		The reader, using any index (auto-detected)
+	 * @return
+	 */
+	public static DatFileReader autoTranslate(int fileIndex, DatFileReader reader, ImageProvider imageProvider) {
+		if (fileIndex == 36) { // use our own ships
+			return new CustomShipsDatFile(reader, imageProvider);
+		}
+		return reader;
+	}
+
 }
