+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/translator/HeaderType.java
@@ -1,5 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2015
+/*
+ * Copyright (c) 2015 - 2018
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -11,30 +11,25 @@
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
- *******************************************************************************/
-package jsettlers.graphics.image;
-
-import jsettlers.common.images.EImageLinkType;
-import jsettlers.graphics.image.reader.ImageMetadata;
+ */
+package jsettlers.graphics.image.reader.translator;
 
 /**
- * This is a gui image.
+ * The type of header that given {@link DatBitmapTranslator} expects.
  * 
  * @author Michael Zangl
- * @see EImageLinkType#GUI
  */
-public class GuiImage extends SingleImage {
-
+public enum HeaderType {
 	/**
-	 * Creates a new GUI image.
-	 * 
-	 * @param metadata
-	 *            The image meta data
-	 * @param data
-	 *            The image data.
+	 * A header that contains a x,y displacement of the image.
 	 */
-	public GuiImage(ImageMetadata metadata, short[] data) {
-		super(metadata, data);
-	}
-
+	DISPLACED,
+	/**
+	 * A special landscape type header.
+	 */
+	LANDSCAPE,
+	/**
+	 * A header with the size for GUI images.
+	 */
+	GUI
 }
