+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/translator/LandscapeTranslator.java
@@ -1,5 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2015
+/*
+ * Copyright (c) 2015 - 2018
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -11,49 +11,39 @@
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
- *******************************************************************************/
-package jsettlers.graphics.reader.translator;
+ */
+package jsettlers.graphics.image.reader.translator;
 
 import java.io.IOException;
 
-import jsettlers.graphics.image.SettlerImage;
-import jsettlers.graphics.reader.DatFileType;
-import jsettlers.graphics.reader.ImageMetadata;
-import jsettlers.graphics.reader.bytereader.ByteReader;
+import jsettlers.graphics.image.LandscapeImage;
+import jsettlers.graphics.image.reader.bytereader.ByteReader;
+import jsettlers.graphics.image.reader.DatFileType;
+import jsettlers.graphics.image.reader.ImageMetadata;
 
 /**
- * This class translates settler images. Settler images are packed in sequences.
- *
+ * This class converts the background landscape images.
+ * 
  * @author Michael Zangl
  *
  */
-public class SettlerTranslator implements DatBitmapTranslator<SettlerImage> {
+public class LandscapeTranslator implements DatBitmapTranslator<LandscapeImage> {
 
 	private final DatFileType type;
 
 	/**
-	 * Creates a new {@link SettlerTranslator}.
+	 * Create a new {@link LandscapeTranslator}.
 	 * 
 	 * @param type
 	 *            The {@link DatFileType} to convert colors.
 	 */
-	public SettlerTranslator(DatFileType type) {
+	public LandscapeTranslator(DatFileType type) {
 		this.type = type;
 	}
 
 	@Override
-	public SettlerImage createImage(ImageMetadata metadata, short[] array) {
-		return new SettlerImage(metadata, array);
-	}
-
-	@Override
 	public short getTransparentColor() {
-		return 0x00;
-	}
-
-	@Override
-	public HeaderType getHeaderType() {
-		return HeaderType.DISPLACED;
+		return 0;
 	}
 
 	@Override
@@ -61,4 +51,13 @@ public class SettlerTranslator implements DatBitmapTranslator<SettlerImage> {
 		return type.convertTo5551(reader.read16());
 	}
 
+	@Override
+	public HeaderType getHeaderType() {
+		return HeaderType.LANDSCAPE;
+	}
+
+	@Override
+	public LandscapeImage createImage(ImageMetadata metadata, short[] array) {
+		return new LandscapeImage(metadata, array);
+	}
 }
