+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/translator/SettlerTranslator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018
+ * Copyright (c) 2015 - 2018
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -12,44 +12,53 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
-package jsettlers.graphics.reader;
+package jsettlers.graphics.image.reader.translator;
 
-import jsettlers.common.images.AnimationSequence;
-import jsettlers.graphics.image.Image;
-import jsettlers.graphics.image.NullImage;
-import jsettlers.graphics.map.draw.ImageProvider;
-import jsettlers.graphics.image.sequence.Sequence;
+import java.io.IOException;
+
+import jsettlers.graphics.image.SettlerImage;
+import jsettlers.graphics.image.reader.bytereader.ByteReader;
+import jsettlers.graphics.image.reader.DatFileType;
+import jsettlers.graphics.image.reader.ImageMetadata;
 
 /**
- * Mapps an {@link AnimationSequence} to the acutal images using an
- * {@link ImageProvider}
- * 
- * @author michael
+ * This class translates settler images. Settler images are packed in sequences.
+ *
+ * @author Michael Zangl
  *
  */
-public class WrappedAnimation implements Sequence<Image> {
+public class SettlerTranslator implements DatBitmapTranslator<SettlerImage> {
 
-	private final ImageProvider imageProvider;
-	private final AnimationSequence sequence;
+	private final DatFileType type;
 
-	public WrappedAnimation(ImageProvider imageProvider, AnimationSequence sequence) {
-		this.imageProvider = imageProvider;
-		this.sequence = sequence;
+	/**
+	 * Creates a new {@link SettlerTranslator}.
+	 * 
+	 * @param type
+	 *            The {@link DatFileType} to convert colors.
+	 */
+	public SettlerTranslator(DatFileType type) {
+		this.type = type;
 	}
 
 	@Override
-	public int length() {
-		return sequence.getLength();
+	public SettlerImage createImage(ImageMetadata metadata, short[] array) {
+		return new SettlerImage(metadata, array);
 	}
 
 	@Override
-	public Image getImage(int index) {
-		return imageProvider.getImage(sequence.getImage(index));
+	public short getTransparentColor() {
+		return 0x00;
 	}
 
 	@Override
-	public Image getImageSafe(int index) {
-		return index < 0 || index >= length() ? NullImage.getInstance() : getImage(index);
+	public HeaderType getHeaderType() {
+		return HeaderType.DISPLACED;
+	}
+
+	@Override
+	public short readUntransparentColor(ByteReader reader) throws IOException {
+		return type.convertTo5551(reader.read16());
 	}
 
 }
