+++ jsettlers.graphics/src/main/java/jsettlers/graphics/map/draw/Background.java
@@ -31,11 +31,10 @@ import jsettlers.common.map.IGraphicsBackgroundListener;
 import jsettlers.common.map.shapes.MapRectangle;
 import jsettlers.common.position.FloatRectangle;
 import jsettlers.graphics.map.MapDrawContext;
-import jsettlers.graphics.reader.AdvancedDatFileReader;
-import jsettlers.graphics.reader.DatBitmapReader;
-import jsettlers.graphics.reader.DatFileReader;
-import jsettlers.graphics.reader.ImageArrayProvider;
-import jsettlers.graphics.reader.ImageMetadata;
+import jsettlers.graphics.image.reader.DatBitmapReader;
+import jsettlers.graphics.image.reader.DatFileReader;
+import jsettlers.graphics.image.reader.ImageArrayProvider;
+import jsettlers.graphics.image.reader.ImageMetadata;
 
 /**
  * The map background.
