+++ jsettlers.graphics/src/main/java/jsettlers/graphics/map/draw/ImageProvider.java
@@ -31,12 +31,12 @@ import jsettlers.graphics.image.ImageIndexFile;
 import jsettlers.graphics.image.LandscapeImage;
 import jsettlers.graphics.image.NullImage;
 import jsettlers.graphics.image.SingleImage;
-import jsettlers.graphics.reader.AdvancedDatFileReader;
-import jsettlers.graphics.reader.custom.graphics.DatFileIndexUtils;
-import jsettlers.graphics.reader.DatFileReader;
-import jsettlers.graphics.reader.DatFileSet;
-import jsettlers.graphics.reader.DatFileType;
-import jsettlers.graphics.reader.EmptyDatFile;
+import jsettlers.graphics.image.reader.AdvancedDatFileReader;
+import jsettlers.graphics.image.reader.custom.graphics.DatFileIndexUtils;
+import jsettlers.graphics.image.reader.DatFileReader;
+import jsettlers.graphics.image.reader.DatFileSet;
+import jsettlers.graphics.image.reader.DatFileType;
+import jsettlers.graphics.image.reader.EmptyDatFile;
 import jsettlers.graphics.image.sequence.ArraySequence;
 import jsettlers.graphics.image.sequence.Sequence;
 import jsettlers.common.images.TextureMap;
