+++ jsettlers.graphics/src/main/java/jsettlers/graphics/sound/SoundManager.java
@@ -26,7 +26,7 @@ import jsettlers.common.map.shapes.MapRectangle;
 import jsettlers.common.position.ShortPoint2D;
 import jsettlers.common.utils.FileUtils;
 import jsettlers.graphics.map.MapDrawContext;
-import jsettlers.graphics.reader.bytereader.ByteReader;
+import jsettlers.graphics.image.reader.bytereader.ByteReader;
 
 /*
  * This class manages reading and playing of the sound file.
