+++ jsettlers.tools/src/main/java/jsettlers/graphics/debug/DatFileMappingComparator.java
@@ -13,9 +13,9 @@ import jsettlers.common.utils.Tuple;
 import jsettlers.graphics.image.GuiImage;
 import jsettlers.graphics.image.Image;
 import jsettlers.graphics.image.SingleImage;
-import jsettlers.graphics.reader.AdvancedDatFileReader;
-import jsettlers.graphics.reader.DatFileType;
-import jsettlers.graphics.reader.SequenceList;
+import jsettlers.graphics.image.reader.AdvancedDatFileReader;
+import jsettlers.graphics.image.reader.DatFileType;
+import jsettlers.graphics.image.sequence.SequenceList;
 import jsettlers.graphics.image.sequence.Sequence;
 
 /**
